requirements_win_10_x64.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
easydict==1.9
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.10
importlib-metadata==4.0.1
Markdown==3.3.4
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboard-pytorch==0.7.1
torch==1.8.1+cu102
torchaudio==0.8.1
torchfile==0.1.0
torchvision==0.9.1+cu102
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
